# Optimize peak_local_max trivial image test