--- a/skimage/feature/peak.py
+++ b/skimage/feature/peak.py
@@ -189,7 +189,7 @@ def peak_local_max(image, min_distance=1, threshold_abs=None,
             "length as the dimensionality of the image.")
 
     # no peak for a trivial image
-    if np.all(image == image.flat[0]):
+    if all((val == image.flat[0] for val in np.nditer(image))):
         if indices is True:
             return np.empty((0, image.ndim), np.int)
         else: