## Description

Closes #3990 

Using a generator to perform the [trivial image check](https://github.com/scikit-image/scikit-image/blob/master/skimage/feature/peak.py#L192) in `peak_local_max` makes the operation O(1) instead of O(n).

Some timings to show the difference:

```python
In [1]: x = np.random.rand(1000, 1000)                                                                                 

In [2]: %timeit np.all(x == x.flat[0])                                                                                 
718 µs ± 1.03 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [3]: %timeit all((v == x.flat[0] for v in np.nditer(x)))                                                            
6.29 µs ± 35.1 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [4]: x = np.ones((1000, 1000))                                                                                      

In [5]: %timeit np.all(x == x.flat[0])                                                                                 
760 µs ± 2.39 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [6]: %timeit all((v == x.flat[0] for v in np.nditer(x)))                                                            
2.3 s ± 11.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

The speedup is more then 100× for a random image!

## Checklist

<!-- It's fine to submit PRs which are a work in progress! -->
<!-- But before they are merged, all PRs should provide: -->
- [ ] [Docstrings for all functions](https://github.com/numpy/numpy/blob/master/doc/example.py)
- [ ] Gallery example in `./doc/examples` (new features only)
- [ ] Benchmark in `./benchmarks`, if your changes aren't covered by an
  existing benchmark
- [ ] Unit tests
- [ ] Clean style in [the spirit of PEP8](https://www.python.org/dev/peps/pep-0008/)

<!-- For detailed information on these and other aspects see -->
<!-- the scikit-image contribution guidelines. -->
<!-- https://scikit-image.org/docs/dev/contribute.html -->

## For reviewers

<!-- Don't remove the checklist below. -->
- [ ] Check that the PR title is short, concise, and will make sense 1 year
  later.
- [ ] Check that new functions are imported in corresponding `__init__.py`.
- [ ] Check that new features, API changes, and deprecations are mentioned in
      `doc/release/release_dev.rst`.