@rfezzani 2020-05-26
I performed some more experiments for performance comparison:

```python
In [8]: for i in range(0, 1000000, 10000): 
   ...:     x = np.ones((1000, 1000)) 
   ...:     x[np.unravel_index(i, (1000, 1000))] = 0 
   ...:     %timeit all((v == x.flat[0] for v in np.nditer(x))) 
```

Compiling the results in a plot gives:
![timings](https://user-images.githubusercontent.com/3438227/82904939-78791000-9f63-11ea-94e0-15128dfff939.png)

The proposed approach only makes sense if we assume that the condition `image[idx] != image[0]` occurs  early when traversing `image` in most of the cases, otherwise, the O(n) approach is the best choice!