@rfezzani 2020-05-28
`peak_local_max` trivial image test is better managed in #4760. This PR is no more relevant, closing ;)